refactor gather_statements
